requirements.txt
chainlit
langchain
faiss-cpu
model2vec
python-dotenv